+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * docs(nuxt.config): reduce component meta bundle size
 * fix(module): detect lazy components when using `experimental.componentDetection`
 * fix(NavigationMenu/Tabs): ensure proper badge display
+* fix(Button): width for one icon
 * test(Form): skip tests because of race condition
 
 ## [2.0.6](https://github.com/bitrix24/b24ui/compare/v2.0.5...v2.0.6) (2025-10-28)